fix(comment): reset github comment on push
Closes #960
