+++ apps/github/src/index.ts
@@ -1,6 +1,7 @@
 import type { components } from "@octokit/openapi-types";
 
 export * from "./client.js";
+export * from "./comment.js";
 
 export type GhApiRepository = components["schemas"]["repository"];
 export type GhApiInstallation = components["schemas"]["installation"];
